Initial Code Folder for Space Adventure.
Haven't added Ships, Trade, Currency, or Planets, but it does provide the initial outline of races, a couple of weapons, and quests. It is playable in its current state though it doesn't do much yet. Give it a try and we will add to the code and storyline and work out the bugs.